logs
add client to log
remove clients from log
search log by date
pagination on left table

missing
- dialog when a client is on the log already
- search input on right table
- pagination right table?
- bug fix when a log is created and there are no clients

+++ backEnd/controllers/bitacoraController.js
@@ -51,7 +51,7 @@ bitacoraCtr.getBitacora = async(req,res) =>{
   try{
       const db = dbConnection.getInstance();
       const bitacora = defineBitacora(db.Sequelize,db.dataType);
-      const bitacora_date = req.params
+      const {bitacora_date} = req.params
       
       let fecha = convertirFechaADate(bitacora_date);
       
@@ -122,37 +122,35 @@ bitacoraCtr.clienteSaveBitacora = async(req,res) =>{
   }
 }
 
-bitacoraCtr.clienteDeleteBitacora = async (req, res) => {
-  try {
+bitacoraCtr.clienteDeleteBitacora = async(req,res) =>{
+  try{
     const db = dbConnection.getInstance();
-    const data = req.body;
-    const defineList = defineClienteXBitacora(db.Sequelize, db.dataType);
-
-    // Usar Promise.all para esperar a que todas las eliminaciones se completen
-    const destroyedValues = await Promise.all(
-      data.deleteList.map(async (item) => {
-        return await defineList.destroy({
-          where: {
-            bitacora_id: data.bitacora_id,
-            cliente_servicio_id: item.cliente_servicio_id,
-          },
-        });
-      })
-    );
-
-    return res.status(200).json({
-      message: "Los usuarios o el usuario se borraron de la bitacora",
-      data: destroyedValues,
-      status: 200,
+    const data = req.body 
+    const defineList =  defineClienteXBitacora(db.Sequelize,db.dataType);
+    console.log(data)
+    const resultFind = await defineList.findOne({
+      where:{
+        bitacora_id:data.bitacora_id,
+        cliente_servicio_id:data.cliente_servicio_id
+      }
     });
-  } catch (error) {
-    console.error("Error al borrar el cliente:", error);
-    res.status(500).json({
-      message: "Error al borrar la información del usuario en la bitacora",
-      status: 500,
-    });
+    if(resultFind){
+      const destroyedValue = await defineList.destroy({
+        where:{
+          bitacora_id:data.bitacora_id,
+          cliente_servicio_id:data.cliente_servicio_id
+        }
+      });
+      return res.status(200).json({message: "El usuario se borro de la bitacora",data:destroyedValue,status:200});
+    }
+    
+    return res.status(404).json({message: "El usuario no existe en la bitacora",status:404});
+
+  }catch(error){
+    console.error("Error al guardar el cliente:", error)
+    res.status(500).json({ message: "Error al obtener la informacion del usuario a buscar en la bitacora",status:500 });
   }
-};
+}
 
 bitacoraCtr.getLastRoom = async(req,res) =>{
   try{
